requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
Brotli==1.1.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.0.1
gunicorn==21.2.0
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.12.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wcwidth==0.2.10
whitenoise==6.6.0
